Move Preferences to top level
